requirements.txt
deepface
numpy
opencv-python
Pillow
PyJWT
PyMuPDF
pymongo
python-dotenv
scikit-learn
torch
torchvision